requirements.txt
notifiers==1.2.1
python-dotenv==0.16.0
venmo-api==0.3.0
gspread==6.1.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.127.0
